back

+++ src/sinlut.h
@@ -5,10 +5,10 @@
 
 extern const s16 sinlut[0x200]; // 4.12 fixed point sin table.
 
-INLINE s16 lu_sin(u16 a)
-{ return sinlut[a>>7]; }
+INLINE s32 lu_sin(u16 a)
+{ return (s32)sinlut[a>>7]; }
 
-INLINE s16 lu_cos(u16 a)
+INLINE s32 lu_cos(u16 a)
 { return lu_sin(a+0x4000); }
 
 #endif
